Update CONFLICTS definitions in categories a* to c*
Update the CONFLICTS definitions of ports in the following categories:

- accessibility
- archivers
- audio
- benchmarks
- biology
- cad
- chinese
- comms
- converters

An attempt has been made to use generic conflicts patterns that do not
have to be updated whenever a new version of a conflicting port is
added to the ports system.

There is a misunderstanding that the port being built/installed has to
be omitted from the conflicts pattern. This is not true - the port
being built is implicitly non-conflicting due to logic in bsd.port.mk.

Approved by:	portmgr (implicit)