editors/libreoffice: work around changed alignment of __cxa_exception
LibreOffice has special detection for the change in alignment and size
of struct cxa_exception, when using libc++abi. However, this updated
alignment also applies to libunwind and upstream libcxxrt, and will soon
apply to our libcxxrt too.

Enable the special detection unconditionally for x86_64 and aarch64, so
libreoffice packages built on 13.0-R (with the old alignment) will
seamlessly work on 13.1-R (which will have the new alignment).

PR:	262008